+++ src/app/hooks/useSpotifyAPI.tsx
@@ -1,0 +1,7 @@
+import 'server-only';
+import { getSpotifyAPIToken } from "../utils/getSpotifyAPIToken"
+
+export const useSpotifyAPIToken = () => {
+    const response = getSpotifyAPIToken();
+    console.log(response)
+}